requirements.txt
aiofiles==0.4.0
aiohttp==3.8.4
aiologger==0.7.0
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==40.0.2
dill==0.3.6
frozenlist==1.3.3
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
lxml==4.9.2
multidict==6.0.4
multiprocess==0.70.14
openai==0.27.4
pathos==0.3.0
pox==0.3.2
ppft==1.7.6.6
pycparser==2.21
PyExecJS==1.5.1
python-dotenv==1.0.0
python-telegram-bot==20.2
requests==2.29.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tqdm==4.65.0
translators==5.7.1
urllib3==1.26.15
yarl==1.9.2
